problem: enable atxi auto-indexing via cli flag
solution: create --atxi.autobuild to enable automatic index building and bookmarking

- refactor atxi models and db handlers to own file in core/atxi.go
- use blockchain receiver for atxi and atxiprogress structs
- creates 'automode' idea to discern whether or not to use bookmarks for
chronological index progress

rel #597
